fix: does not use db inline updates
Previously, inline updates (called ON DUPLICATE
KEY UPDATE) were used. However, this updates not
only if there is a duplicate primary key but also
if there are duplicate unique keys, that is not
intentional. Now a preceding query is used to
check whether the entity already exists and
either an insert or an update statement is
executed accordingly.